+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Settings page update check backed by the daemon `/updates/check` endpoint.
+
 ## [0.1.0-beta] - 2026-05-14
 
 ### Added